repo_scanner/requirements.txt
boto3==1.18.0
GitPython==3.1.30